+++ src/lib.rs
@@ -1,7 +1,7 @@
 #![doc(html_root_url = "https://docs.rs/rfc2047-decoder/0.1.3")]
 
-pub mod decoder;
-pub use decoder::Decoder;
+mod decoder;
+pub use decoder::{Decoder, Error, Result};
 
 mod evaluator;
 mod lexer;
@@ -9,6 +9,6 @@ mod parser;
 
 /// Decodes the given RFC 2047 MIME Message Header encoded string
 /// using a default decoder.
-pub fn decode<T: AsRef<[u8]>>(encoded_str: T) -> decoder::Result<String> {
+pub fn decode<T: AsRef<[u8]>>(encoded_str: T) -> Result<String> {
     Decoder::new().decode(encoded_str)
 }